feat:update model traininig on windows OS

+++ train.py
@@ -2,8 +2,8 @@ import os
 import sys
 import pprint
 import random
-import logging
 import argparse
+import platform
 from pathlib import Path
 from copy import deepcopy
 from datetime import datetime
@@ -20,6 +20,7 @@ from torch.nn.parallel import DistributedDataParallel as DDP
 SEED = 2023
 random.seed(SEED)
 torch.manual_seed(SEED)
+OS_SYSTEM = platform.system()
 ROOT = Path(__file__).resolve().parents[0]
 TIMESTAMP = datetime.today().strftime("%Y-%m-%d_%H-%M")
 
@@ -27,7 +28,7 @@ from dataloader import build_dataset
 from model import build_model
 from utils import (build_optimizer, build_criterion, build_scheduler,
                     set_lr, setup_worker_logging, setup_primary_logging, 
-                    resume_state, de_parallel)
+                    build_logger, resume_state, de_parallel)
 from val import validate
 
 
@@ -38,15 +39,17 @@ def setup(rank, world_size):
         rank (int): distributed process ID
         world_size (int): umber of distributed processes
     """
-    os.environ['MASTER_ADDR'] = 'localhost'
-    os.environ['MASTER_PORT'] = '5555'
-    dist.init_process_group('nccl', rank=rank, world_size=world_size)
+    if OS_SYSTEM == "Linux":
+        os.environ["MASTER_ADDR"] = "localhost"
+        os.environ["MASTER_PORT"] = "2355"
+        dist.init_process_group("nccl", rank=rank, world_size=world_size)
 
 
 def cleanup():
     """close DDP backend communication.
     """
-    dist.destroy_process_group()
+    if OS_SYSTEM == "Linux":
+        dist.destroy_process_group()
 
 
 def train(args, dataloader, model, criterion, optimizer, scaler):
@@ -70,11 +73,11 @@ def train(args, dataloader, model, criterion, optimizer, scaler):
     for i, minibatch in enumerate(dataloader):
         ni = i + len(dataloader) * (epoch - 1)
         if ni <= args.nw:
-            set_lr(optimizer, args.base_lr * pow(ni / (args.nw), 4))
+            set_lr(optimizer, args.base_lr * pow(ni / (args.nw), 2))
 
         images, labels = minibatch[0], minibatch[1]
         with amp.autocast(enabled=not args.no_amp):
-            predictions = model.forward(images.cuda(args.rank, non_blocking=True))['out']
+            predictions = model(images.cuda(args.rank, non_blocking=True))
             loss = criterion(predictions, labels.cuda(args.rank, non_blocking=True))
 
         scaler.scale(loss * args.world_size).backward()
@@ -107,6 +110,8 @@ def parse_args(make_dirs=True):
     parser.add_argument('--exp', type=str, required=True, help='Name to log training')
     parser.add_argument('--data', type=str, default='mit67', help='Dataset name(must be match to <dataset>.yaml')
     parser.add_argument('--model', type=str, default='resnet18', help='Model architecture')
+    parser.add_argument('--mobile-v3', type=str, default='large', help='Mobilenetv3 architecture mode')
+    parser.add_argument('--img-size', type=int, default=224, help="Model input size")
     parser.add_argument('--loss', type=str, default='ce', help='Loss function')
     parser.add_argument('--optim', type=str, default='sgd', help='Optimizor for training')
     parser.add_argument('--batch-size', type=int, default=256, help='Batch size')
@@ -118,10 +123,13 @@ def parse_args(make_dirs=True):
     parser.add_argument('--weight-decay', type=float, default=1e-4, help='Weight decay')
     parser.add_argument('--label-smoothing', type=float, default=0.0, help='Label smoothing')
     parser.add_argument('--workers', type=int, default=8, help='Number of workers for dataloader')
-    parser.add_argument('--no-amp', action='store_true', help='Use of FP32 training (default: AMP training)')
-    parser.add_argument('--resume', action='store_true', help='Name to resume path')
-    parser.add_argument('--rank', type=int, default=0, help='Process id for computation')
     parser.add_argument('--world-size', type=int, default=1, help='Number of available GPU devices')
+    parser.add_argument('--rank', type=int, default=0, help='Process id for computation')
+    parser.add_argument('--no-amp', action='store_true', help='Use of FP32 training (default: AMP training)')
+    parser.add_argument('--width-multiple', type=float, default=1.0, help='CSP-Layer channel multiple')
+    parser.add_argument('--depth-multiple', type=float, default=1.0, help='CSP-Model depth multiple')
+    parser.add_argument('--pretrained', action='store_true', help='Training with pretrained weights')
+    parser.add_argument('--resume', action='store_true', help='Name to resume path')
     
     args = parser.parse_args()
     args.data = ROOT / 'data' / args.data
@@ -147,7 +155,12 @@ def main_work(rank, world_size, args, logger):
     ################################### Init Process ####################################
     setup(rank, world_size)
     torch.cuda.set_device(rank)
-    setup_worker_logging(rank, logger)
+
+    if OS_SYSTEM == "Linux":
+        import logging
+        setup_worker_logging(rank, logger)
+    else:
+        logging = logger
 
     ################################### Init Instance ###################################
     global epoch
@@ -159,17 +172,15 @@ def main_work(rank, world_size, args, logger):
     args.workers = min([os.cpu_count() // max(world_size, 1), args.batch_size if args.batch_size > 1 else 0, args.workers])
 
     ################################### Init Instance ###################################
-    train_dataset, val_dataset, hyp = build_dataset(yaml_path=str(args.data)+'.yaml')
+    train_dataset, val_dataset, hyp = build_dataset(yaml_path=str(args.data)+'.yaml', input_size=args.img_size)
     train_sampler = distributed.DistributedSampler(dataset=train_dataset, num_replicas=world_size, rank=args.rank, shuffle=True)
     train_loader = DataLoader(dataset=train_dataset, batch_size=args.batch_size, shuffle=False, 
                               pin_memory=True, num_workers=args.workers, sampler=train_sampler)
-    val_loader = DataLoader(dataset=val_dataset, batch_size=args.batch_size, 
-                            shuffle=False, pin_memory=True, num_workers=args.workers)
+    val_loader = DataLoader(dataset=val_dataset, batch_size=args.batch_size, shuffle=False, pin_memory=True, num_workers=args.workers)
     args.nw = max(round(args.warmup * len(train_loader)), 100)
-    model = build_model(arch_name=args.arch, num_classes=len(hyp['CLASS_INFO']), width_multiple=args.width_multiple, 
+    model = build_model(arch_name=args.model, num_classes=len(hyp['CLASS_INFO']), width_multiple=args.width_multiple, 
                         depth_multiple=args.depth_multiple, mode=args.mobile_v3, pretrained=args.pretrained)
-    macs, params = profile(deepcopy(model), inputs=(torch.randn(1, 3, hyp['INPUT_SIZE'], hyp['INPUT_SIZE']),), verbose=False)
-    
+    macs, params = profile(deepcopy(model), inputs=(torch.randn(1, 3, args.img_size, args.img_size),), verbose=False)
     criterion = build_criterion(name=args.loss, label_smoothing=args.label_smoothing)
     optimizer = build_optimizer(model=model, name=args.optim, lr=args.base_lr, momentum=args.momentum, weight_decay=args.weight_decay)
     scheduler = build_scheduler(optimizer=optimizer, name=args.data.name, num_epoch=args.num_epoch, lr_decay=args.lr_decay)
@@ -187,8 +198,9 @@ def main_work(rank, world_size, args, logger):
             logging.warning(f'Arch info - Params(M): {params/1e+6:.2f}, FLOPs(G): {2*macs/1E+9:.2f}')
 
     #################################### Train Model ####################################
-    model = DDP(model, device_ids=[args.rank])
-    dist.barrier()
+    if OS_SYSTEM == "Linux":
+        model = DDP(model, device_ids=[args.rank])
+        dist.barrier()
     
     if args.rank == 0:
         pbar = trange(start_epoch, args.num_epoch+1, total=args.num_epoch, initial=start_epoch, ncols=110)
@@ -199,17 +211,21 @@ def main_work(rank, world_size, args, logger):
     for epoch in pbar:
         if args.rank == 0:
             train_loader = tqdm(train_loader, desc=f'[TRAIN:{epoch:03d}/{args.num_epoch:03d}]', ncols=110, leave=False)
+        
         train_sampler.set_epoch(epoch)
         train_loss_msg = train(args=args, dataloader=train_loader, model=model, criterion=criterion, optimizer=optimizer, scaler=scaler)
         
         if args.rank == 0:
-            save_obj = {'current_epoch': epoch, 
-                        'idx2cls': hyp['CLASS_INFO'],
-                        'arch_name': args.arch, 
-                        'model_state': deepcopy(de_parallel(model)).state_dict(),
-                        'optimizer_state': optimizer.state_dict(),
-                        'scheduler_state': scheduler.state_dict(),
-                        'scaler_state_dict': scaler.state_dict()}
+            save_obj = {"running_epoch": epoch,
+                        "class_list": hyp['CLASS_INFO'],
+                        "model": args.model,
+                        "mobile_v3": args.mobile_v3, 
+                        "width_multiple": args.width_multiple,
+                        "depth_multiple": args.depth_multiple,
+                        "model_state": deepcopy(de_parallel(model)).state_dict(),
+                        "optimizer_state": optimizer.state_dict(),
+                        "scheduler_state": scheduler.state_dict(),
+                        "scaler_state_dict": scaler.state_dict()}
             torch.save(save_obj, args.weight_dir / 'last.pt')
 
             val_loader = tqdm(val_loader, desc=f'[VAL:{epoch:03d}/{args.num_epoch:03d}]', ncols=110, leave=False)
@@ -219,6 +235,7 @@ def main_work(rank, world_size, args, logger):
             if top1_acc > best_score:
                 best_epoch, best_score, best_perf_msg = epoch, top1_acc, eval_msg
                 torch.save(save_obj, args.weight_dir / 'best.pt')
+        
         scheduler.step()
 
     if args.rank == 0:
@@ -228,6 +245,11 @@ def main_work(rank, world_size, args, logger):
 
 if __name__ == "__main__":
     args = parse_args(make_dirs=True)
-    torch.multiprocessing.set_start_method('spawn', force=True)
-    logger = setup_primary_logging(args.exp_path / 'train.log')
-    mp.spawn(main_work, args=(args.world_size, args, logger), nprocs=args.world_size, join=True)
+    
+    if OS_SYSTEM == "Linux":
+        torch.multiprocessing.set_start_method('spawn', force=True)
+        logger = setup_primary_logging(args.exp_path / 'train.log')
+        mp.spawn(main_work, args=(args.world_size, args, logger), nprocs=args.world_size, join=True)
+    else:
+        logger = build_logger(args.exp_path / 'train.log')
+        main_work(rank=0, world_size=1, args=args, logger=logger)
